update :: pages widget name and page name changed
