+++ lib/main.dart
@@ -1,16 +1,16 @@
 import 'package:flutter/material.dart';
 
-import 'package:resumecraft/pages/profile/profile_section.dart';
+import 'package:resumecraft/pages/profile/profile_section_page.dart';
 import 'package:resumecraft/pages/auth/login_page.dart';
 import 'package:resumecraft/pages/home/home_page.dart';
-import 'package:resumecraft/pages/profile_section/education.dart';
-import 'package:resumecraft/pages/profile_section/experience.dart';
-import 'package:resumecraft/pages/profile_section/objective.dart';
-import 'package:resumecraft/pages/profile_section/personal_detail.dart';
+import 'package:resumecraft/pages/profile_section/education_page.dart';
+import 'package:resumecraft/pages/profile_section/experience_page.dart';
+import 'package:resumecraft/pages/profile_section/objective_page.dart';
+import 'package:resumecraft/pages/profile_section/personal_detail_page.dart';
 import 'package:resumecraft/pages/profile/profile_page.dart';
 import 'package:resumecraft/pages/auth/register_page.dart';
-import 'package:resumecraft/pages/profile_section/project.dart';
-import 'package:resumecraft/pages/profile_section/skill.dart';
+import 'package:resumecraft/pages/profile_section/project_page.dart';
+import 'package:resumecraft/pages/profile_section/skill_page.dart';
 import 'package:resumecraft/utils/shared_prefs/user_shared_prefs.dart';
 
 Widget _defaultHome = const LoginPage(); // Default to LoginPage initially
@@ -58,15 +58,15 @@ class MyApp extends StatelessWidget {
         '/profiles': (context) => const ProfilePage(),
 
         // ================= profile-sections ===================
-        '/profile-section': (context) => const ProfileSection(),
+        '/profile-section': (context) => const ProfileSectionPage(),
 
         // ================= create-section-detail ===================
-        '/personal-detail': (context) => const PersonalDetail(),
-        '/education': (context) => const Education(),
-        '/experience': (context) => const Experience(),
-        '/skill': (context) => const Skill(),
-        '/project': (context) => const Project(),
-        '/objective': (context) => const Objective(),
+        '/personal-detail': (context) => const PersonalDetailPage(),
+        '/education': (context) => const EducationPage(),
+        '/experience': (context) => const ExperiencePage(),
+        '/skill': (context) => const SkillPage(),
+        '/project': (context) => const ProjectPage(),
+        '/objective': (context) => const ObjectivePage(),
       },
     );
   }
