+++ lib/pages/profile_section/project_page.dart
@@ -2,31 +2,34 @@ import 'package:flutter/material.dart';
 import 'package:snippet_coder_utils/FormHelper.dart';
 import 'package:snippet_coder_utils/hex_color.dart';
 
-import 'package:resumecraft/services/objective_api_service.dart';
-import 'package:resumecraft/utils/mixins/Objective/objective_mixin.dart';
-import 'package:resumecraft/models/profile_section/objective/write/objective_request_model.dart';
+import 'package:resumecraft/services/project_api_service.dart';
+import 'package:resumecraft/utils/mixins/project/project_mixin.dart';
+import 'package:resumecraft/models/profile_section/projects/write/project_request_model.dart';
 import 'package:resumecraft/utils/mixins/user/user_mixin.dart';
 
 import '../../config.dart';
 
-class Objective extends StatefulWidget {
-  const Objective({super.key});
+class ProjectPage extends StatefulWidget {
+  const ProjectPage({super.key});
 
   @override
-  State<Objective> createState() => _ObjectiveState();
+  State<ProjectPage> createState() => _ProjectPageState();
 }
 
-class _ObjectiveState extends State<Objective>
-    with UserProfileMixin, ObjectiveMixin {
+class _ProjectPageState extends State<ProjectPage>
+    with UserProfileMixin, ProjectMixin {
   final Color primaryColor = HexColor('#283B71');
   final GlobalKey<FormState> globalFormKey = GlobalKey<FormState>();
   bool isApicallProcess = false;
-  String? details;
+  String? projectTitle;
+  String? projectDesc;
+  String? projLinks;
+  List<String> links = [];
 
   Widget build(BuildContext context) {
     final args =
         ModalRoute.of(context)?.settings.arguments as Map<String, dynamic>?;
-    final id = args?['ObjectiveId'] as String?;
+    final id = args?['ProjectId'] as String?;
 
     if (id != null) {
       setPersonalDetailId(id);
@@ -55,23 +58,72 @@ class _ObjectiveState extends State<Objective>
           children: [
             FormHelper.inputFieldWidgetWithLabel(
               context,
-              "details",
-              "Details",
+              "projectTitle",
+              "Project Title",
               "",
               (onValidateVal) {
                 if (onValidateVal.isEmpty) {
-                  return 'Objective details cannot be empty';
+                  return 'Project Title cannot be empty';
                 }
                 return null;
               },
               (onSavedVal) {
-                details = onSavedVal;
+                projectTitle = onSavedVal;
               },
-              initialValue: objective?.details ?? "",
+              initialValue: project?.projectTitle ?? "",
               borderFocusColor: primaryColor,
               borderColor: primaryColor,
               textColor: Colors.black,
               hintColor: Colors.black.withOpacity(0.7),
+              borderRadius: 10,
+              labelFontSize: 16,
+              paddingLeft: 0,
+              paddingRight: 0,
+            ),
+            const SizedBox(height: 16),
+            FormHelper.inputFieldWidgetWithLabel(
+              context,
+              "projectDesc",
+              "Project Description",
+              "",
+              (onValidateVal) {
+                if (onValidateVal.isEmpty) {
+                  return 'Project Description cannot be empty';
+                }
+                return null;
+              },
+              (onSavedVal) {
+                projectDesc = onSavedVal;
+              },
+              initialValue: project?.projectDesc ?? "",
+              borderFocusColor: primaryColor,
+              borderColor: primaryColor,
+              textColor: Colors.black,
+              hintColor: Colors.black.withOpacity(0.7),
+              borderRadius: 10,
+              labelFontSize: 16,
+              paddingLeft: 0,
+              paddingRight: 0,
+            ),
+            FormHelper.inputFieldWidgetWithLabel(
+              context,
+              "links",
+              "Links",
+              "www.this.com, www.that.com, ...",
+              (onValidateVal) {
+                if (onValidateVal.isEmpty) {
+                  return 'At least one project link should be an input';
+                }
+                return null;
+              },
+              (onSavedVal) {
+                projLinks = onSavedVal;
+              },
+              initialValue: (project?.links)?.join(',') ?? links.join(','),
+              hintColor: Colors.black45,
+              borderFocusColor: primaryColor,
+              borderColor: primaryColor,
+              textColor: Colors.black,
               borderRadius: 10,
               labelFontSize: 16,
               paddingLeft: 0,
@@ -88,20 +140,22 @@ class _ObjectiveState extends State<Objective>
                     });
 
                     try {
-                      ObjectiveRequestModel model = ObjectiveRequestModel(
+                      ProjectRequestModel model = ProjectRequestModel(
                         userdetail: personalDetailID,
-                        details: details,
+                        projectTitle: projectTitle!,
+                        projectDesc: projectDesc!,
+                        links: links,
                       );
-                      if (objective != null) {
-                        final response =
-                            await ObjectiveAPIService.updateObjective(
-                                model, userToken, personalDetailID);
+                      if (project != null) {
+                        final response = await ProjectAPIService.updateProject(
+                            model, userToken, personalDetailID);
                         setState(() {
                           isApicallProcess = false;
                         });
                         if (response.statusCode == 200) {
-                          FormHelper.showSimpleAlertDialog(context,
-                              Config.appName, "Objective edited!", "OK", () {
+                          FormHelper.showSimpleAlertDialog(
+                              context, Config.appName, "Project edited!", "OK",
+                              () {
                             Navigator.pop(context);
                             Navigator.pushReplacementNamed(
                                 context, '/profile-section');
@@ -110,14 +164,13 @@ class _ObjectiveState extends State<Objective>
                           FormHelper.showSimpleAlertDialog(
                               context,
                               Config.appName,
-                              "Failed to edit Objective. Please try again.",
+                              "Failed to edit project. Please try again.",
                               "OK",
                               () => Navigator.pop(context));
                         }
                       } else {
-                        final response =
-                            await ObjectiveAPIService.createObjective(
-                                model, userToken);
+                        final response = await ProjectAPIService.createProject(
+                            model, userToken);
                         setState(() {
                           isApicallProcess = false;
                         });
@@ -133,7 +186,7 @@ class _ObjectiveState extends State<Objective>
                           FormHelper.showSimpleAlertDialog(
                               context,
                               Config.appName,
-                              "Failed to save Objective. Please try again.",
+                              "Failed to save project. Please try again.",
                               "OK",
                               () => Navigator.pop(context));
                         }
@@ -170,6 +223,9 @@ class _ObjectiveState extends State<Objective>
     final form = globalFormKey.currentState;
     if (form!.validate()) {
       form.save();
+      if (projLinks?.isNotEmpty ?? false) {
+        links = projLinks!.split(',').map((link) => link.trim()).toList();
+      }
       return true;
     } else {
       return false;
